Embedded/requirement.txt
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorzero==2.0
cryptography==43.0.1
gcloud==0.18.3
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
gpiozero==2.0.1
grpcio==1.66.1
grpcio-status==1.66.1
gTTS==2.5.3
httplib2==0.22.0
idna==3.10
importlib-metadata==4.13.0
importlib-resources==5.13.0
jwcrypto==1.5.6
numpy==2.0.2
oauth2client==4.1.3
opencv-python==4.10.0.84
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1-modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.1
pydantic-core==2.23.3
pygame==2.6.0
pynmea2==1.19.0
pyparsing==3.1.4
Pyrebase4==4.8.0
pyserial==3.5
python-jwt==4.1.0
requests==2.32.3
requests-toolbelt==0.10.1
RPi.GPIO==0.7.1
rsa==4.9
six==1.16.0
tqdm==4.66.5
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
zipp==3.20.2
